+++ io/TIndexReader.cpp
@@ -41,6 +41,7 @@
 #include <pdal/util/ProgramArgs.hpp>
 #include <pdal/private/gdal/GDALUtils.hpp>
 #include <pdal/private/gdal/SpatialRef.hpp>
+#include <pdal/util/private/JsonSupport.hpp>
 #include <pdal/StageWrapper.hpp>
 
 #include <nlohmann/json.hpp>
@@ -78,112 +79,6 @@ struct TIndexReader::Args
     std::vector<NL::json> m_rawReaderArgs;
     NL::json m_readerArgs;
 };
-
-namespace
-{
-
-NL::json handleReaderArgs(NL::json rawReaderArgs)
-{
-    if (rawReaderArgs.is_object())
-    {
-        NL::json array_args = NL::json::array();
-        array_args.push_back(rawReaderArgs);
-        rawReaderArgs = array_args;
-    }
-    for (auto& opts: rawReaderArgs)
-        if (!opts.is_object())
-            throw pdal_error("Reader Args for each reader must be a valid JSON object");
-
-    NL::json readerArgs;
-    for (NL::json& readerPipeline: rawReaderArgs)
-    {
-
-        if (!readerPipeline.contains("type"))
-            throw pdal_error("No \"type\" key found in supplied reader arguments.");
-
-        std::string driver = readerPipeline.at("type").get<std::string>();
-        if (rawReaderArgs.contains(driver))
-            throw pdal_error("Multiple instances of the same driver in supplied reader arguments.");
-        readerArgs[driver] = { };
-
-        for (auto& arg: readerPipeline.items())
-        {
-            if (arg.key() == "type")
-                continue;
-
-            std::string key = arg.key();
-            readerArgs[driver][key] = { };
-            readerArgs[driver][key] = arg.value();
-        }
-    }
-    return readerArgs;
-}
-
-Options setReaderOptions(const NL::json& readerArgs, const std::string& driver,
-    const std::string& filename)
-{
-    Options readerOptions;
-    bool filenameSet = false;
-    if (readerArgs.contains(driver)) {
-        NL::json args = readerArgs.at(driver).get<NL::json>();
-        for (auto& arg : args.items()) {
-            if (arg.key() == "filename")
-            {
-                NL::json filespecArg = arg.value().get<NL::json>();
-                if (!filespecArg.is_object())
-                    throw pdal_error("Value for " + driver + " 'filename' argument " +
-                        "expected to be a 'FileSpec' JSON object.");
-                if (filespecArg.contains("path"))
-                    filespecArg.erase("path");
-                filespecArg += {"path", filename};
-
-                // This doesn't check if the driver supports headers/queries: if not,
-                // the reader will only use the filename
-                readerOptions.add("filename", filespecArg.dump());
-                filenameSet = true;
-                continue;
-            }
-
-            NL::detail::value_t type = readerArgs.at(driver).at(arg.key()).type();
-            switch(type)
-            {
-                case NL::detail::value_t::string:
-                {
-                    std::string val = arg.value().get<std::string>();
-                    readerOptions.add(arg.key(), arg.value().get<std::string>());
-                    break;
-                }
-                case NL::detail::value_t::number_float:
-                {
-                    readerOptions.add(arg.key(), arg.value().get<float>());
-                    break;
-                }
-                case NL::detail::value_t::number_integer:
-                {
-                    readerOptions.add(arg.key(), arg.value().get<int>());
-                    break;
-                }
-                case NL::detail::value_t::boolean:
-                {
-                    readerOptions.add(arg.key(), arg.value().get<bool>());
-                    break;
-                }
-                default:
-                {
-                    readerOptions.add(arg.key(), arg.value());
-                    break;
-                }
-            }
-        }
-    }
-    if (!filenameSet)
-        readerOptions.add("filename", filename);
-
-
-    return readerOptions;
-}
-
-} // unnamed namespace
 
 TIndexReader::TIndexReader() :
     m_args(new TIndexReader::Args),
@@ -363,7 +258,7 @@ void TIndexReader::initialize()
 
 
     if (m_args->m_rawReaderArgs.size())
-        m_args->m_readerArgs = handleReaderArgs(m_args->m_rawReaderArgs);
+        m_args->m_readerArgs = Utils::handleReaderArgs(m_args->m_rawReaderArgs);
 
     for (auto f : getFiles())
     {
@@ -377,7 +272,7 @@ void TIndexReader::initialize()
                 "'.");
         reader->setLog(log());
 
-        Options readerOptions = setReaderOptions(m_args->m_readerArgs, driver, f.m_filename);
+        Options readerOptions = Utils::setReaderOptions(m_args->m_readerArgs, driver, f.m_filename);
 
         reader->setOptions(readerOptions);
         Stage *premerge = reader;
